Separation of OSD specific code into a shared library
All OSD and vicep-access related code is compiled into two
different shared libraries: libafsosd.so and libdafsosd.so, the
second one for dafs servers. The fileserver and volserver got a new
parameter "-libafsosd". With this parameter they load and initialize
the shared library. During the initialization operation vectors
are filled which allow the code inside the library to use programs
in the binary and vice versa. The binaries contain a number of hooks
into the shared library which anly will be called when the shared
library is loaded. Thus the same binary can be used for "normal"
fileservers and OSD-fileservers. The salvager also loads the library
when it detects an OSD-volume.

All OSD or vicep-access specific RPCs are separated into a secondary service
which is started when "-libafsosd" was present. However, for backward
compatibility in this version all the old RPCs are still in place, but
are handled in the shared library rather than in the binary. In the openafs
git master these compatibility will hopefully not be required any more.

Some additional goodies which are not OSD specific such as the RPCs
needed for the "fs threads"-command, the "fs statistic"-command and
the "fs getvariable"-, "fs setvariable"-command I kept in the main
binary.

The StartAsyncFetch-, ExtendAsyncFetch-, EndAsyncFetch-, StartAsyncStore-,
ExtendAsyncStore, and EndAsyncStore-RPCs remained in th main binary, but
got where necessary opaque parameters which are unmarshalled and marshalled
in the shared library. This construct allows future use of the same RPCs also
in the main binary.

The interface contains a version number which must be the same in the binary and
the shared library to prevent loading an incompatible library. This version
number must be incremented whenever the structs ore opearation parameters
of the interface change.

The commands (fs, vos,...) are in this version only partially converted
to be linked against a libafsosd.a library. Also the cache manager is not yet
broken up into two separate modules. This fill follow...

The fileserver is already running on one server in our cell and in the
test cell on my laptop, but I am sure it is not yet production quality.

I removed all the man-pages from git which were created by "regen.sh" and
should not be stored in git.