Use offset + length in EndAsyncStore
EndAsyncStore used to use only the file length reported by the
client to calculate the new length of the file. When the store
is split in multiple transactions because the client doesn't
store more than 16 chunks at a time this length can be larger
than the actual length stored on the OSD. Therefore the same
calculation as StoreData_RXStyle is applied here.