Allow more than 64 chunks at a time to be written to OSDs
My understanding is that the limit to store only 64 chunks at a time
to the fileserver should help to keep RPC times low. For OSDs this
limit seems unnecessary so I allow here 256 chunks at a time. This
reduces the overhead of StartAsyncStore and EndAsyncStore RPCs and
lets the data version number grow less fast.