Allow ubik to handle multiple databases
In preparation for letting the vlserver also serve the OSDDB I opened ubik to
handle more than one database. The number of possible databases is given
by a #define for MAX_UBIK_DBASES which currently is 4. The ubik services
VOTE and DISK have been modified on the wire, but as far as I know in a cell
all databse servers of a sort should run the same software release, anyway.
So this protocoll change should not lead to incompatibilities.

udebug is still backward compatible to the old ubik, but shows per database
status in the present version.

The sync site is be the same for all databases. So any additional database
must be available on the sync site, but the additional databases
don't need to be started on all non-sync-site servers (leading then to a
recoverystate of only f for these databases).

There is a new call ubik_ServerInitByInfoN to start the Nth database. For easy
backward compatibility the old ubik_ServerInitByInfo is still supported for
the primary database.