Integration of OSDD into vlserver
As Jeffrey Altman proposed in Hamburg in October the OSDDB should
run as an additional service in the vlserver. There are only few
additions in vlserver.c to start the osddb from libafsosd.so.

Since libafsosd.so is compiled for AFS_PTHREAD_ENV this code is
included only in the vlserver under src/tvlserver. When the
correct version of libafsosd.so is installed and the vlserver
is started with the parameter "-libafsosd" then it serves also
the OSDDB under its original service number 13. For backward
compatibility with our openafs-1.4-osd environment ths OSDDB
listens also on port 7012.

There airnot yet any changes on the client side to support port 7003
instead of 7012 in this version.