Function isOsdFile and better versioning
New function isOsdFile called by osdvol->op_isOsdFile to check
whether a file is on local disk or in OSD. This is called in the
fileserver, volserver, and salvager.

I also made the version number of afsosd.h become the new subversion
number of the shared libraries. Presently we have libafsosd.0.6 and
libdafsosd.0.6 (for DAFS servers).