Do more when setting or resetting osdPolicy
If the volume is a normal openafs volume create the osdMetdataFile
and reset vnode magic which now will be used as osdMetadataIndex.

When resetting it to zero make sure there are no more OSD files.
Otherwise give up. In future the volume could be reset to the
state of an normal openafs volume recreating magics and deleting
the osdMetadataFile, but this is still ifdefed to remain
compatible with openafs-1.4.x-osd.