Separate HPSS stuff into a shared library
All source which needs hpss include files gets compiled in the new
subdirectoriy src/shlibafshpss and is not build by default. This
is necessary to allow public rpms to be built automatically. Sites
using hpss and having the appropriate licenses may build separately
the shared library libafshpss.so.

Now the rxosd always uses ih_ops in the ihandle to access i/o routines.
This makes the code some what simpler.

The same rxosd binary can be used with or without hpss depending on
the existence of the shared library and of at least the -hpss_meta
parameter. The hpss root directory can now be specified in the
HPSS.conf file as well as the principal and the keytab file to be
used for authentication.

The same changes have been made for dcache. However the subdirectory
src/shlibafsdcache is still missing (because I presently have no access
to dcache in order to test my code).