Correct bugs in the dump/restore processing
There was still a problem when an OSD volume was replicated the
1st time to another server: the setting of osdPolicy was called
incorrectly with VolumeDiskData instead of Volume as parameter.
Also one operation had been forgotten to be filled in vol_osd.c.
Another bug was that the length of OSD-files was not transmitted
to the sink site ending up all OSD-files with length zero.