Let libafsosd fileserver announce this cabability
and allow the client only open connections to service VICEOSD_SERVICE
if the server has announced its capability. The problem here is that
packets to an unknown service are simply discarded. Then the client
sets the server down and resets it up once the normal service (1)
is reachable again.

Also some code cleanup in afs_vicep.c