New primary port 7006, but listen also on 7012
In 1.6 and the master the rxosd port should be 7006. But
we need for compatibility with 1.4 also port 7012.

Also use port and service from OSDDB entry if available when
starting a connection to another rxosd.

Allow to specify the basic port as start parameter -port.
Allow to speicify the number of lwps by -lwps (debugging
is easier with only a few lwps!)