procs in volserosd.xg must be in ascending opcode order
volserosd.ss.c gets a jump-table which contains entries in the order
the proc statements appear in the source .xg file. Therefore the
order is important.

Also the Variable rpc was missing which can be used to change LogLevel and
convertToOsd. it has been added to both volserosd.xg and volserosd.c