Use PIO also in rxosd
There was something strange when HAVE_PIO was switched off: too many
file-descriptors remained open. For HPSS we simulate pread and pwrite
by a sequence of lseek and read or lseek and write which may not be
absolutely thread-save, but there shouldn't be concurrent data streams
for the same file, anyway, in an archival rxosd.