Avoid unlinked files when link table is in bad state
in incdec():

If link count is zero check whether file really exists.
If not return EIO;
If it exists and we should do a decrement call namei_SetNonZLC to
correct the linktable. It sets a very high link.

Use in SRXOSD_builincdec and SRXOSD_bulkinc152 also incdec() to
have the same behaviour.