Build fs and vos witch shared library
in the rxosd directory a version fo vos and fs is build which requires
the shared libary libcafsosd.so.0.<num> where <num> is presently 21.
This library must be installed in a public place like /lib64. If the
library cannot be found fs and vos behave like in standard openafs.

The advantage is that the number of changes in the venus and volser
directories can be reduced. Right now, however, the old source is still
there and youcan build fs and vos also in the classical directories as
before. The new "stripped" source files are called fs_t.c and vos_t.c,
the "t" because they are to be built pthreaded.