Changes by Achim Gsell
In order to get rid of warnings and also because many format statements
where wrong and some parameters used pointers to objects instead of the
objects themselves Achim proposed many of these changes. There are also
changes in favour of DARWIN platform. PrintTime and PrintDate now take
the time as afs_uint32 instead of a pointer to afs_uint32.