Fix for a dangerous bug in vol_split.c
If the volume being split has osdPolicy == 0, but has an osdMetadataHandle then
in the new created volume the vnodes didn't get the field magic filled.
If, however, the volume was created freshly on a 1.6 fileserver or moved
to it the volume to be split doesn't have an osdMetadataHandle and the vnodes
had already the correct magic. In the 1st case the salvager wouldn't load the
libafsosd extensions and would delete all the vnodes without correct magic!

Please remember: you shouldn't start an openafs-osd-1.6 fileserver for
vicep-partitions used before in an openafs-osd-1.4 fileserver. If you start using
openafs-osd-1.6 do it with empty partitions and fill them by moving the volumes
there (or creating new volumes).