Avoid fetch entry reference count to stay
osd rmfetchentry returns 16 (busy) if the reference count of the fetch entry
isn't zero. Now at rxosd restart, when the fetchqueue is read in from disk, the
refcnt is reset. Also in XferData should now be no way to exit without
decrementing the reference count. I hope that will help...