Let archival disk OSDs return 'p' for HSM status
Before wiping a file the fileserver checks the HSM state to be either
'p' == premigrated or 'm' == migrated. Up to now archival disk OSDs used
to be secondary archival copies, so the problem didn't occur because the
other copy on a real HSM system showed the 'p' or 'm'. In order to make
this happen easily a new struct myOsd was introduced which contains the
fields from the OSDDB.