Dangerous bug!!
in AllocMetadataEntry the last patch by Achim changed the way
the end of the alloc table entry was calculated to point 40 bytes
after the end of the alloc table entry. This resulted in never
allocating the next alloc table entry writing all further
entries to the same place.

The old calculation is now back here.