Allow again "vos convertROtoRW" for osd volumes
With taking openafs's original namei_ops.c in the vol-subdirectory
this feature was broken because the volserver didn't know how to
handle the osdMetadata volume spaecial file. Now after a vos convert
the osdMetadata file is removed when the volume had no osdPolicy
otherwise it is kept and the also in the volume header file
correctly denoted.