Keep OsdMetadata file even if osdPolicy == 0
During "vos convertROtoRW" processing we should not remove the
OsdMatadata file for volumes without osdPolicy because the vnode files
don't contain the vnode magic numbers. Without the Osdmetadata file the
salvager would not load libafsosd and check the vnode magic numbers. This
would result in deleting all files and directories.