Make openafs-1.6-osd compatible to version for openafs master
The version for openafs master doesn't support any of the additional
RPCs in volint.xg and afsint.xg we still support here for
1.4-osd servers and clients.

Before this patch when moving or releasing a volume a RPC
AFSVolOsdSupport was used to find out whether the target side
understands AFS/OSD. Now only for move AFSVOLOSD_OsdSupport is called
while release to a non-AFS/OSD server is not possible any more
for OSD-volumes. It also wouldn't make much sense!

Other changes are:

The osddbserver now uses libafsosd.so.0.xx instead of containing the
code of osddbprocs.c itself.

afsio in our AFS/OSD version is compiled in subdirectory src/rxosd
while in src/venus the openafs afsio which is based on libafscp and
does not understand AFS/OSD is compiled.

vos and fs are now only installed from src/rxosd. These are pthreaded and
load libcafsosd.so.0.xx.