Close newly created file with ih_ops->close routine
for files created or opened in HPSS to avoid tumbling of fd numbers the fds get
by a constant. During file creation it was tried to close them by OS_CLOSE which
is not correct. This could lead to too many open files or to closing a local file
instead of the HPSS file.