upgrade to 1.8.0pre5 and some small changes
To the small changes belongs that the ubik database servers got an option
"-haveAlwaysSyncSite". This option can be used in small cells where not always all
machines wich database servers on them are up and the network consists in a single
switch or router which is supposed to be stable. When set even a single machine running
can be sync-site.